Now we are using the real data
Used datafrmaes to import data.
The value of currene(current energy) was allowed to vary from the current choosen value (uniform distribution with witdth 13.6*2 ev). This was done because for a fixed value of energy the discreteness led to huge variability in percentage formation, just by changing the energy by 1ev in a direction.