Was using wrong value of elastic cross section
Little need to use anything else but the average value, because even after huge variations done in the averaged value of thresholdel the percentage of positronium remained almost same. 
Analyze why: One obvious thing observed is that the temperature , density and value of elastic cross section effect the percentage of pos formation via the value of alpha parameter.